# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title": "Iron Feather", "version": "0.1.0", "description": "OGC API Features Core and GeoJSON. Read-only CRS84 features from a DuckDB shard (local file or read-only HTTP/S3 attachment). XYZ MVT is an extension, not an OGC API Tiles conformance claim."},
  "servers": [{"url": "/"}],
  "paths": {
    "/": {"get": {"operationId": "landing", "responses": {"400": {"$ref": "#/components/responses/BadRequest"}, "200": {"description": "Landing page", "content": {"application/json": {"schema": {"type": "object", "required": ["links"], "properties": {"title": {"type": "string"}, "description": {"type": "string"}, "links": {"$ref": "#/components/schemas/Links"}}}}}}}}},
    "/conformance": {"get": {"operationId": "conformance", "responses": {"400": {"$ref": "#/components/responses/BadRequest"}, "200": {"description": "Conformance classes", "content": {"application/json": {"schema": {"type": "object", "required": ["conformsTo"], "properties": {"conformsTo": {"type": "array", "items": {"type": "string", "format": "uri"}}}}}}}}}},
    "/api": {"get": {"operationId": "api", "responses": {"400": {"$ref": "#/components/responses/BadRequest"}, "200": {"description": "This OpenAPI document", "content": {"application/vnd.oai.openapi+json;version=3.0": {"schema": {"type": "object"}}}}}}},
    "/api.html": {"get": {"operationId": "documentation", "responses": {"400": {"$ref": "#/components/responses/BadRequest"}, "200": {"description": "HTML API documentation", "content": {"text/html": {"schema": {"type": "string"}}}}}}},
    "/healthz": {"get": {"operationId": "health", "responses": {"400": {"$ref": "#/components/responses/BadRequest"}, "200": {"description": "Ready", "content": {"text/plain": {"schema": {"type": "string"}}}}}}},
    "/collections": {"get": {"operationId": "collections", "responses": {"400": {"$ref": "#/components/responses/BadRequest"}, "200": {"description": "Collections present in the shard", "content": {"application/json": {"schema": {"type": "object", "required": ["links", "collections"], "properties": {"links": {"$ref": "#/components/schemas/Links"}, "collections": {"type": "array", "items": {"$ref": "#/components/schemas/Collection"}}}}}}}}}},
    "/collections/{collectionId}": {
      "parameters": [{"$ref": "#/components/parameters/collectionId"}],
      "get": {"operationId": "collection", "responses": {"400": {"$ref": "#/components/responses/BadRequest"}, "200": {"description": "Collection metadata", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Collection"}}}}, "404": {"$ref": "#/components/responses/Error"}}}
    },
    "/collections/{collectionId}/items": {
      "parameters": [{"$ref": "#/components/parameters/collectionId"}, {"$ref": "#/components/parameters/sources"}, {"$ref": "#/components/parameters/sourceHeader"}],
      "get": {
        "operationId": "items",
        "description": "Stable id-ordered pagination. numberMatched is omitted. Unknown parameters (including filter and properties) are rejected. No sources means no visible rows.",
        "parameters": [
          {"name": "bbox", "in": "query", "description": "CRS84 west,south,east,north, or west,south,minHeight,east,north,maxHeight. West greater than east crosses the antimeridian. Height does not restrict 2D features.", "style": "form", "explode": false, "schema": {"type": "array", "minItems": 4, "maxItems": 6, "items": {"type": "number"}}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 1000, "default": 10}},
          {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0, "maximum": 4294967295, "default": 0}},
          {"name": "datetime", "in": "query", "style": "form", "explode": false, "description": "RFC3339 instant or start/end interval, with an open endpoint written as .. or empty. Layercake edit timestamps are not temporal geometry; these static features match every valid datetime.", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "GeoJSON feature page", "content": {"application/geo+json": {"schema": {"$ref": "#/components/schemas/FeatureCollection"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/Error"}, "406": {"description": "Unsupported Accept media type"}, "429": {"$ref": "#/components/responses/Overloaded"}, "500": {"$ref": "#/components/responses/Error"}
        }
      }
    },
    "/collections/{collectionId}/items/{featureId}": {
      "parameters": [{"$ref": "#/components/parameters/collectionId"}, {"name": "featureId", "in": "path", "required": true, "schema": {"type": "string"}}, {"$ref": "#/components/parameters/sources"}, {"$ref": "#/components/parameters/sourceHeader"}],
      "get": {"operationId": "item", "responses": {
        "200": {"description": "GeoJSON feature", "content": {"application/geo+json": {"schema": {"$ref": "#/components/schemas/Feature"}}}},
        "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/Error"}, "406": {"description": "Unsupported Accept media type"}, "429": {"$ref": "#/components/responses/Overloaded"}, "500": {"$ref": "#/components/responses/Error"}
      }}
    },
    "/collections/{collectionId}/tiles/{z}/{x}/{y}": {
      "parameters": [{"$ref": "#/components/parameters/collectionId"}, {"$ref": "#/components/parameters/sources"}, {"$ref": "#/components/parameters/sourceHeader"},
        {"name": "z", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 0, "maximum": 30}},
        {"name": "x", "in": "path", "required": true, "description": "Less than 2^z", "schema": {"type": "integer", "minimum": 0}},
        {"name": "y", "in": "path", "required": true, "description": "Less than 2^z", "schema": {"type": "integer", "minimum": 0}}],
      "get": {"operationId": "tile", "description": "Web Mercator XYZ MVT; extent 4096, at most 5000 features.", "responses": {
        "200": {"description": "Vector tile", "content": {"application/vnd.mapbox-vector-tile": {"schema": {"type": "string", "format": "binary"}}}},
        "204": {"description": "Empty tile"}, "400": {"$ref": "#/components/responses/BadRequest"}, "404": {"$ref": "#/components/responses/Error"}, "429": {"$ref": "#/components/responses/Overloaded"}, "500": {"$ref": "#/components/responses/Error"}
      }}
    }
  },
  "components": {
    "parameters": {
      "collectionId": {"name": "collectionId", "in": "path", "required": true, "schema": {"type": "string"}},
      "sources": {"name": "sources", "in": "query", "description": "Comma-separated source IDs; intersected with X-Source-Ids when both are supplied. Data filter, not authentication. Missing both yields no rows.", "style": "form", "explode": false, "schema": {"type": "array", "items": {"type": "integer", "format": "int64"}}},
      "sourceHeader": {"name": "X-Source-Ids", "in": "header", "description": "Comma-separated source IDs", "schema": {"type": "string"}}
    },
    "responses": {
      "BadRequest": {"description": "Invalid or unsupported request parameter", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}, "text/plain": {"schema": {"type": "string"}}}},
      "Error": {"description": "Request failed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "Overloaded": {"description": "All DuckDB connections are busy", "headers": {"Retry-After": {"schema": {"type": "integer", "enum": [1]}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
    },
    "schemas": {
      "Error": {"type": "object", "required": ["code", "description"], "properties": {"code": {"type": "integer"}, "description": {"type": "string"}}},
      "Link": {"type": "object", "required": ["href", "rel"], "properties": {"href": {"type": "string", "format": "uri-reference"}, "rel": {"type": "string"}, "type": {"type": "string"}, "title": {"type": "string"}}},
      "Links": {"type": "array", "items": {"$ref": "#/components/schemas/Link"}},
      "Collection": {"type": "object", "required": ["id", "links"], "properties": {"id": {"type": "string"}, "title": {"type": "string"}, "itemType": {"type": "string", "enum": ["feature"]}, "links": {"$ref": "#/components/schemas/Links"}}},
      "Geometry": {"type": "object", "nullable": true, "required": ["type"], "properties": {"type": {"type": "string", "enum": ["Point", "MultiPoint", "LineString", "MultiLineString", "Polygon", "MultiPolygon", "GeometryCollection"]}, "coordinates": {"type": "array", "items": {}}, "geometries": {"type": "array", "items": {"$ref": "#/components/schemas/Geometry"}}}},
      "Feature": {"type": "object", "required": ["type", "id", "geometry", "properties"], "properties": {"type": {"type": "string", "enum": ["Feature"]}, "id": {"type": "string"}, "geometry": {"$ref": "#/components/schemas/Geometry"}, "properties": {"type": "object", "nullable": true, "additionalProperties": true}, "links": {"$ref": "#/components/schemas/Links"}}},
      "FeatureCollection": {"type": "object", "required": ["type", "features", "links", "numberReturned"], "properties": {"type": {"type": "string", "enum": ["FeatureCollection"]}, "features": {"type": "array", "items": {"$ref": "#/components/schemas/Feature"}}, "links": {"$ref": "#/components/schemas/Links"}, "numberReturned": {"type": "integer", "minimum": 0}}}
    }
  }
}